requirements.txt
# =============================================================================
# COMICS CLOZE VLM - Python Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# Note: For GPU training, install PyTorch with CUDA first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# =============================================================================

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Transformers & PEFT (LoRA)
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0
bitsandbytes>=0.43.0

# Dataset & Data Processing
datasets>=2.18.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Evaluation Metrics
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13
evaluate>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter & Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0

# Cloud & Storage
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.36.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jsonlines>=4.0.0
rich>=13.0.0

# Optional: OCR (if preprocessing from scratch)
# pytesseract>=0.3.10
# easyocr>=1.7.0

# Optional: Streamlit Web Interface
# streamlit>=1.28.0
# streamlit-chat>=0.1.1
